pyproject.toml
[project]
name = "mcp-ecobici"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.1",
    "fastapi[all]>=0.115.13",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mcp[cli]>=1.9.4",
    "pandas>=2.3.0",
    "pydantic-ai>=0.3.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "sqlmodel>=0.0.24",
]
